ADD: create new file for handleLogin method
