+++ model/users.json
@@ -1,1 +1,8 @@
-[]
+[
+    {
+        "id": "685fb30c-4533-4dfe-bd0d-d35de604cf11",
+        "username": "wayncis",
+        "password": "$2b$10$Zj.5.aSoNklG0BEx/aN1AeUp.Nl3aKH7mzlVo1ShE6K6FECvZ1VZW"
+    }
+      
+]
